configured useRoute to set a uuid and for apply to check uuid before wrapping resolvers, also cleaned up console logs and comments
